Fixes after checking CRC
