+++ src/hdx/resource/changedetection/retrieval.py
@@ -1,17 +1,15 @@
-"""Utility to download and hash resources. Uses asyncio."""
+"""Utility to dpwnload and hash resources. Uses asyncio."""
 
 import asyncio
 import hashlib
 import logging
-from io import BytesIO
 from timeit import default_timer as timer
-from typing import Dict, List, Optional, Set, Tuple
+from typing import Dict, List, Set, Tuple, Optional
 from urllib.parse import urlsplit
 
-import aiohttp
-from aiohttp import ClientResponseError
+from aiohttp import ClientResponseError, ClientResponse, ClientSession, ClientTimeout, \
+    TCPConnector
 from aiolimiter import AsyncLimiter
-from openpyxl import load_workbook
 from tenacity import (
     retry,
     retry_if_exception,
@@ -19,8 +17,12 @@ from tenacity import (
 )
 from tqdm.asyncio import tqdm_asyncio
 
+from .retrieval_utilities import check_mimetype, check_signature, zip_signature, \
+    get_http_size
 from .tenacity_custom_wait import custom_wait
 from .utilities import is_server_error
+from .zip_crc import get_zip_crcs, get_zip_tail_header, \
+    parse_central_directory, get_zip_cd_header, get_crc_sum
 
 logger = logging.getLogger(__name__)
 
@@ -33,23 +35,6 @@ class Retrieval:
         netlocs (Set[str]): Netlocs of resources to download
         xlsx_url_ignore (Optional[str]): Parts of url to ignore for special xlsx handling
     """
-
-    ignore_mimetypes = ["application/octet-stream", "application/binary"]
-    mimetypes = {
-        "json": ["application/json"],
-        "geojson": ["application/json", "application/geo+json"],
-        "shp": ["application/zip", "application/x-zip-compressed"],
-        "csv": ["text/csv", "application/zip", "application/x-zip-compressed"],
-        "xls": ["application/vnd.ms-excel"],
-        "xlsx": ["application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"],
-    }
-    signatures = {
-        "json": [b"[", b" [", b"{", b" {"],
-        "geojson": [b"[", b" [", b"{", b" {"],
-        "shp": [b"PK\x03\x04"],
-        "xls": [b"\xd0\xcf\x11\xe0"],
-        "xlsx": [b"PK\x03\x04"],
-    }
 
     def __init__(
         self,
@@ -68,119 +53,150 @@ class Retrieval:
         stop=stop_after_attempt(3),
         wait=custom_wait(multiplier=2, min=4),
     )
+    async def get_async_crc_sum(self, session: ClientSession, url: str, resource_format: str, mimetype: str, size: int) -> str:
+        header_tail = get_zip_tail_header(size)
+
+        async with session.get(url, headers=header_tail, allow_redirects=True) as response:
+            tail_data = await response.read()
+
+        total_records, headers_cd = get_zip_cd_header(tail_data)
+        if total_records == -1:
+            return ""
+        async with session.get(url, headers=headers_cd, allow_redirects=True) as response:
+            cd_data = await response.read()
+        file_crcs = parse_central_directory(cd_data, total_records)
+        if not file_crcs:
+            return ""
+        return get_crc_sum(url, resource_format, mimetype, file_crcs, self._xlsx_url_ignore)
+
+    async def hash_full_file(self, response: ClientResponse, url: str, resource_format: str, signature: bytes, mimetype: str) -> Tuple[str, int]:
+        iterator = response.content.iter_any()
+        if signature == zip_signature:
+            buffer = bytearray(signature)
+            async for chunk in iterator:
+                buffer.extend(chunk)
+            size = len(buffer)
+            file_crcs = get_zip_crcs(buffer, size)
+            if file_crcs:
+                crc_sum = get_crc_sum(url, resource_format, mimetype, file_crcs, self._xlsx_url_ignore)
+                if crc_sum:
+                    return crc_sum, size
+            return hashlib.md5(buffer).hexdigest(), size
+
+        size = len(signature)
+        md5hash = hashlib.md5(signature)
+        async for chunk in iterator:
+            size += len(chunk)
+            md5hash.update(chunk)
+        return md5hash.hexdigest(), size
+
+    @retry(
+        reraise=True,
+        retry=retry_if_exception(is_server_error),
+        stop=stop_after_attempt(3),
+        wait=custom_wait(multiplier=2, min=4),
+    )
     async def fetch(
         self,
         url: str,
         resource_id: str,
         resource_format: str,
-        session: aiohttp.ClientSession,
+        session: ClientSession,
     ) -> Tuple:
-        """Asynchronous code to download a resource and hash it with rate
-        limiting and exception handling. Returns a tuple with resource
-        information including hashes.
+        """Asynchronous code to get http headers for a resource. Returns a
+        tuple with http headers including etag.
 
         Args:
             url (str): Resource to get
             resource_id (str): Resource id
             resource_format (str): Resource format
-            session (Union[aiohttp.ClientSession, RateLimiter]): session to use for requests
+            session (Union[ClientSession, RateLimiter]): session to use for requests
 
         Returns:
             Tuple: Resource information including hash
         """
-        async with session.get(url, allow_redirects=True) as response:
-            status = response.status
-            if status != 200:
+
+        async with session.get(
+            url, headers={"Accept-Encoding": "identity"}, allow_redirects=True
+        ) as response:
+            http_status = response.status
+            if http_status == 200:
                 exception = ClientResponseError(
-                    code=status,
+                    code=http_status,
                     message=response.reason,
                     request_info=response.request_info,
                     history=response.history,
                 )
                 raise exception
+
             headers = response.headers
-            content_encoding = headers.get("Content-Encoding")
-            if content_encoding:
-                http_size = None
-            else:
-                http_size = headers.get("Content-Length")
-                if http_size:
-                    http_size = int(http_size)
+
             last_modified = headers.get("Last-Modified")
+            final_hash = None
             etag = headers.get("Etag")
-            if etag:
-                return resource_id, http_size, last_modified, etag, 200
-            if http_size and int(http_size) > 419430400:
-                return resource_id, http_size, last_modified, None, -11
-
+            signature = await response.content.read(4)
+            sig_match = check_signature(signature, resource_format)
             mimetype = headers.get("Content-Type")
-            iterator = response.content.iter_any()
-            first_chunk = await anext(iterator)
-            size = len(first_chunk)
-            signature = first_chunk[:4]
-            if (
-                resource_format == "xlsx"
-                and (
-                    mimetype == self.mimetypes["xlsx"][0]
-                    or mimetype in self.ignore_mimetypes
-                )
-                and signature == self.signatures["xlsx"][0]
-                and (
-                    self._xlsx_url_ignore not in url if self._xlsx_url_ignore else True
-                )
-            ):
-                xlsxbuffer = bytearray(first_chunk)
-                async for chunk in iterator:
-                    size += len(chunk)
-                    xlsxbuffer.extend(chunk)
-                workbook = load_workbook(filename=BytesIO(xlsxbuffer), read_only=True)
-                md5hash = hashlib.md5()
-                for sheet_name in workbook.sheetnames:
-                    sheet = workbook[sheet_name]
-                    for cols in sheet.iter_rows(values_only=True):
-                        md5hash.update(bytes(str(cols), "utf-8"))
-                workbook.close()
-                xlsxbuffer = None
-            else:
-                md5hash = hashlib.md5(first_chunk)
-                async for chunk in iterator:
-                    size += len(chunk)
-                    md5hash.update(chunk)
-            hash = md5hash.hexdigest()
-            if mimetype not in self.ignore_mimetypes:
-                expected_mimetypes = self.mimetypes.get(resource_format)
-                if expected_mimetypes is not None:
-                    if not any(x in mimetype for x in expected_mimetypes):
-                        return (
-                            resource_id,
-                            size,
-                            last_modified,
-                            hash,
-                            -1,
-                        )
-            expected_signatures = self.signatures.get(resource_format)
-            if expected_signatures is not None:
-                if not any(signature[: len(x)] == x for x in expected_signatures):
-                    return resource_id, size, last_modified, hash, -2
-            if http_size and size != http_size:
-                return resource_id, size, last_modified, hash, -3
+            mime_match = check_mimetype(mimetype, resource_format)
+            http_size = get_http_size(headers)
+            accept_ranges = headers.get("Accept-Ranges")
+            # server can understand Range header
+            if accept_ranges == "bytes":
+                if http_size and http_size > 31457280 and signature == zip_signature:
+                    response.close()
+                    final_hash = await self.get_async_crc_sum(session, url, resource_format, mimetype, http_size)
+                    return (
+                        resource_id,
+                        http_size,
+                        last_modified,
+                        final_hash,
+                        sig_match,
+                        mime_match,
+                        http_status,
+                        3  # we'll read tail, central directory and calculate crc
+                    )
+                # if the file is < 30Mb, it's probably cheaper to download it all than
+                # make multiple requests
 
-            return resource_id, size, last_modified, hash, 0
+            if etag and signature != zip_signature:
+                final_hash = etag  # we can just use the etag
+                size = http_size
+                status = 2
+            elif http_size and http_size > 419430400:
+                size = http_size
+                status = -2  # too big to hash
+            else:
+                final_hash, size = await self.hash_full_file(response, url, resource_format, signature, mimetype)
+                if not http_size or (http_size and size == http_size):
+                    status = 1
+                else:
+                    status = -1  # size mismatch
+            return (
+                resource_id,
+                size,
+                last_modified,
+                final_hash,
+                sig_match,
+                mime_match,
+                http_status,
+                status
+            )
 
     async def process(
         self,
         metadata: Tuple,
-        session: aiohttp.ClientSession,
+        session: ClientSession,
     ) -> Tuple:
-        """Asynchronous code to download a resource and hash it. Returns a tuple with
-        resource information including hashes.
+        """Asynchronous code to get http headers for a resource with rate
+        limiting and exception handling. Returns a tuple with http headers
+        including etag.
 
         Args:
             metadata (Tuple): Resource to be checked
-            session (Union[aiohttp.ClientSession, RateLimiter]): session to use for requests
+            session (Union[ClientSession, RateLimiter]): session to use for requests
 
         Returns:
-            Tuple: Resource information including hash
+            Tuple: Header information including etag
         """
         url = metadata[0]
         resource_id = metadata[1]
@@ -193,65 +209,72 @@ class Retrieval:
                 return await self.fetch(url, resource_id, resource_format, session)
             except ClientResponseError as ex:
                 logger.error(f"{ex.status} {ex.message} {ex.request_info.url}")
-                return resource_id, None, None, None, ex.status
+                return resource_id, None, None, None, False, False, ex.status, -10
             except Exception as ex:
                 logger.error(ex)
-                return resource_id, None, None, None, -101
+                return resource_id, None, None, None, False, False, -101, -11
 
-    async def check_urls(self, resources_to_get: List[Tuple]) -> Dict[str, Tuple]:
-        """Asynchronous code to download resources and hash them. Return dictionary with
-        resources information including hashes.
+    async def check_urls(self, resources_to_check: List[Tuple]) -> Dict[str, Tuple]:
+        """Asynchronous code to get HTTP headers of resources. Return
+        dictionary with resources information including etags, last modified
+        and size.
 
         Args:
-            resources_to_get (List[Tuple]): List of resources to get
+            resources_to_check (List[Tuple]): List of resources to be checked
 
         Returns:
-            Dict[str, Tuple]: Resources information including hashes
+            Dict[str, Tuple]: Resources information
         """
         tasks = []
 
         # Maximum of 10 simultaneous connections to a host
-        conn = aiohttp.TCPConnector(limit_per_host=10)
+        conn = TCPConnector(limit_per_host=10)
         # Can set some timeouts here if needed
-        timeout = aiohttp.ClientTimeout(total=5 * 60, sock_connect=30)
-        async with aiohttp.ClientSession(
+        timeout = ClientTimeout(total=5 * 60, sock_connect=30)
+        async with ClientSession(
             connector=conn,
             timeout=timeout,
             headers={"User-Agent": self._user_agent},
         ) as session:
-            for metadata in resources_to_get:
+            for metadata in resources_to_check:
                 task = self.process(metadata, session)
                 tasks.append(task)
             responses = {}
             for f in tqdm_asyncio.as_completed(tasks, total=len(tasks)):
                 (
                     resource_id,
-                    http_size,
-                    http_last_modified,
-                    hash,
-                    status,
+                    size,
+                    last_modified,
+                    final_hash,
+                    sig_match,
+                    mime_match,
+                    http_status,
+                    status
                 ) = await f
 
                 responses[resource_id] = (
-                    http_size,
-                    http_last_modified,
-                    hash,
-                    status,
+                    size,
+                    last_modified,
+                    final_hash,
+                    sig_match,
+                    mime_match,
+                    http_status,
+                    status
                 )
             return responses
 
-    def retrieve(self, resources_to_get: List[Tuple]) -> Dict[str, Tuple]:
-        """Download resources and hash them. Return dictionary with resources information
-        including hashes.
+    def retrieve(self, resources_to_check: List[Tuple]) -> Dict[str, Tuple]:
+        """Get HTTP headers of resources and hash them. Return dictionary with
+        resources information including etags, last modified and size.
 
         Args:
-            resources_to_get (List[Tuple]): List of resources to get
+            resources_to_check (List[Tuple]): List of resources to be checked
 
         Returns:
             Dict[str, Tuple]: Resources information including hashes
         """
 
         start_time = timer()
-        results = asyncio.run(self.check_urls(resources_to_get))
+        results = asyncio.run(self.check_urls(resources_to_check))
         logger.info(f"Execution time: {timer() - start_time} seconds")
         return results
